refactor(components/atom/helpText): add import file extensions
