+++ components/atom/helpText/src/settings.js
@@ -1,0 +1,1 @@
+export const BASE_CLASS = 'sui-AtomHelpText'
